+++ packages/design-system-web/src/components/geov-entity/geov-entity.stories.tsx
@@ -3,7 +3,7 @@
 import React from 'react';
 import { JSX } from '../..';
 import { GeovEntity } from '../../../.storybook/stencil-generated/component';
-import { DEFAULT_SPARQL_ENDPOINT } from '../../../.storybook/config/defaulSparqlEndpoint';
+import { AMPI_SPARQL_ENDPOINT, DEFAULT_SPARQL_ENDPOINT } from '../../../.storybook/config/defaulSparqlEndpoint';
 
 export default {
   title: 'Data Components/Entity/Entity',
@@ -37,3 +37,48 @@ const args3: JSX.GeovEntity = {
   fetchBeforeRender: false,
 };
 JohannesKepler.args = args3;
+
+export const Birth = Template.bind({});
+const args4: JSX.GeovEntity = {
+  sparqlEndpoint: AMPI_SPARQL_ENDPOINT,
+  entityId: 'i542181',
+  language: 'de',
+  fetchBeforeRender: false,
+};
+Birth.args = args4;
+
+export const TimeSpan = Template.bind({});
+const args5: JSX.GeovEntity = {
+  sparqlEndpoint: AMPI_SPARQL_ENDPOINT,
+  entityId: 'i542181ts',
+  language: 'de',
+  fetchBeforeRender: false,
+};
+TimeSpan.args = args5;
+
+export const DateTimeYear = Template.bind({});
+const arg6: JSX.GeovEntity = {
+  sparqlEndpoint: AMPI_SPARQL_ENDPOINT,
+  entityId: 'i609119',
+  language: 'de',
+  fetchBeforeRender: false,
+};
+DateTimeYear.args = arg6;
+
+export const DateTimeMonth = Template.bind({});
+const arg7: JSX.GeovEntity = {
+  sparqlEndpoint: AMPI_SPARQL_ENDPOINT,
+  entityId: 'i986993',
+  language: 'de',
+  fetchBeforeRender: false,
+};
+DateTimeMonth.args = arg7;
+
+export const DateTimeDay = Template.bind({});
+const arg8: JSX.GeovEntity = {
+  sparqlEndpoint: AMPI_SPARQL_ENDPOINT,
+  entityId: 'i872676',
+  language: 'de',
+  fetchBeforeRender: false,
+};
+DateTimeDay.args = arg8;